Fixes from code review and additional test
