+++ build.gradle
@@ -138,7 +138,7 @@ dependencies {
     generatedCompile 'io.springfox:springfox-swagger2:2.7.0'
     generatedCompile 'io.springfox:springfox-swagger-ui:2.7.0'
 
-    developmentOnly("org.springframework.boot:spring-boot-devtools:${springBootVersion}")
+    developmentOnly("org.springframework.boot:spring-boot-devtools")
 }
 
 def openapiSourceFile = 'src/main/resources/data-repository-openapi.yaml'
